Add(parser): All the final funcs to .h and normm
