+++ parser/ms_parser.c
@@ -6,7 +6,7 @@
 /*   By: gsaiago <gsaiago@student.42.rio>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/28 11:42:31 by gsaiago           #+#    #+#             */
-/*   Updated: 2023/03/19 18:19:20 by gsaiago          ###   ########.fr       */
+/*   Updated: 2023/03/19 19:59:57 by gsaiago          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,8 +26,6 @@ int	ms_parser(char *line, t_word **word_lst)
 		return (1);
 	if (ms_clean_words_and_init_flags_on_lst(word_lst))
 		return (1);
-	//ms_word_lst_flag_init(*word_lst);
-	//ms_lst_remove_if(word_lst);
 	free(line);
 	return (0);
 }
@@ -43,7 +41,7 @@ int	ms_clean_words_and_init_flags_on_lst(t_word **lst)
 	{
 		if (node->word)
 		{
-			//node->flag = ms_flag_word(node); a flag vem antes de retirar as aspas
+			node->flag = ms_flag_word(node);
 			node->word = ms_expand_env(node->word);
 			node->word = ms_remove_quotes(node->word, 1);
 		}
@@ -56,7 +54,6 @@ int	ms_clean_words_and_init_flags_on_lst(t_word **lst)
 int	ms_flag_word(t_word *node)
 {
 	char	*word;
-	int		flag;
 
 	if (!node)
 		return (-1);
@@ -65,12 +62,11 @@ int	ms_flag_word(t_word *node)
 		return (-1);
 	if (ft_strncmp(node->word, "|", 2) == 0)
 		return (MS_PIPE);
-	flag = 1;
-	while (*word)
-	{
-
-	}
-	if (ft_strchr(node->word, '>'))
+	else if (ft_strncmp(node->word, "<<", 3) == 0)
+		return (MS_HEREDOC);
+	else if (ft_strncmp(node->word, ">>", 3) == 0)
+		return (MS_APPEND);
+	else if (ft_strncmp(node->word, ">", 2) == 0)
 		return (MS_REDIRECT);
 	return (MS_WORD);
 }
